refactor(apply-edit): split 626-line apply-proof-edit.ts into focused modules
apply-proof-edit.ts mixed four distinct concerns under one filename:
hardened I/O primitives, goal-keyed edits, batch hole replacements,
and free-form text edits. Each one wants its own test file and reads
better in isolation. Split into:

  safe-source-io.ts      (173) — `MAX_AGDA_SOURCE_BYTES`,
                                 `AgdaSourceReadError`,
                                 `readAgdaSourceFile` (O_NOFOLLOW + size
                                 cap), `loadSourceForEdit`,
                                 `writeFileAtomic` (temp + rename).
  apply-goal-edit.ts     (158) — `applyProofEdit` plus the
                                 `ProofEdit` / `ApplyEditResult` types.
  apply-batch-edits.ts   (147) — `applyBatchHoleReplacements` plus
                                 `BatchApplyResult`.
  apply-text-edit.ts     (191) — `applyTextEdit` plus `TextEditResult`.
  apply-proof-edit.ts     (37) — barrel re-exporting the public API
                                 so existing imports in
                                 `register-agda-apply-edit.ts`,
                                 `reload-and-diagnose.ts`, and the
                                 four test files keep working.

The split preserves the public surface byte-for-byte; 1186 vitest
cases still pass identically. Each new file is well under the 500-line
ceiling so subsequent additions land in the right home rather than
re-inflating one module.
